Change NameContainsKeywordsPredicate to ContainsKeywordsPredicate class
* to accept both name keyword and ic keyword
